Updated the issue number to #497
