+++ processor/src/test/java/org/mapstruct/ap/test/bugs/_497/DoubleMultiplier.java
@@ -1,4 +1,4 @@
-package org.mapstruct.ap.test.bugs._new;
+package org.mapstruct.ap.test.bugs._497;
 
 import org.mapstruct.Qualifier;
 
@@ -9,5 +9,5 @@ import java.lang.annotation.RetentionPolicy;
 @Qualifier
 @java.lang.annotation.Target(ElementType.METHOD)
 @Retention(RetentionPolicy.SOURCE)
-public @interface RandomMultiplier {
+public @interface DoubleMultiplier {
 }
