+++ processor/src/test/java/org/mapstruct/ap/test/bugs/_497/RandomMultiplier.java
@@ -1,4 +1,4 @@
-package org.mapstruct.ap.test.bugs._new;
+package org.mapstruct.ap.test.bugs._497;
 
 import org.mapstruct.Qualifier;
 
@@ -7,7 +7,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
 @Qualifier
-@java.lang.annotation.Target(ElementType.TYPE)
+@java.lang.annotation.Target(ElementType.METHOD)
 @Retention(RetentionPolicy.SOURCE)
-public @interface SourceTypeQualifier {
+public @interface RandomMultiplier {
 }
