+++ processor/src/test/java/org/mapstruct/ap/test/bugs/_497/UseQualifierToDetermineAmbiguityInMappingMethods.java
@@ -16,25 +16,23 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.mapstruct.ap.test.bugs._new;
+package org.mapstruct.ap.test.bugs._497;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mapstruct.ap.testutil.IssueKey;
-import org.mapstruct.ap.testutil.WithClasses;
-import org.mapstruct.ap.testutil.runner.AnnotationProcessorTestRunner;
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.Mappings;
+import org.mapstruct.factory.Mappers;
 
-@WithClasses( {
-    UseQualifierToDetermineAmbiguityInMappingMethods.class,
-    MultiplierMapper.class,
-    Source.class, Target.class, SourceType.class,
-    SourceTypeQualifier.class, RandomMultiplier.class, DoubleMultiplier.class
-} )
-@IssueKey("_new")
-@RunWith(AnnotationProcessorTestRunner.class)
-public class UseQualifierToDetermineAmbiguityInMappingMethodsTest {
+@Mapper(uses = {MultiplierMapper.class})
+public interface UseQualifierToDetermineAmbiguityInMappingMethods {
 
-    @Test
-    public void shouldResolveMethodBasedOnQualifier() {
-    }
+    UseQualifierToDetermineAmbiguityInMappingMethods INSTANCE 
+            = Mappers.getMapper(UseQualifierToDetermineAmbiguityInMappingMethods.class);
+
+    @Mappings({
+            @Mapping(target = "name", source = "name"),
+            @Mapping(target = "multiplier", source = "type",
+                    qualifiedBy = {SourceTypeQualifier.class, RandomMultiplier.class})
+    })
+    Target map(Source source);
 }
